builder: split out builder into a dedicated crate
Split out builder into a dedicate nydus-builder crate, to reduce
dependencies of the nydus-rafs crate.
