+++ Cargo.toml
@@ -57,7 +57,8 @@ openssl = { version = "0.10.48", features = ["vendored"] }
 #openssl-src = { version = "111.22" }
 
 nydus-api = { version = "0.2.2", path = "api", features = ["handler"] }
-nydus-rafs = { version = "0.2.2", path = "rafs", features = ["builder"] }
+nydus-builder = { version = "0.1.0", path = "builder" }
+nydus-rafs = { version = "0.3.0", path = "rafs" }
 nydus-service = { version = "0.2.0", path = "service", features = ["block-device"] }
 nydus-storage = { version = "0.6.2", path = "storage" }
 nydus-utils = { version = "0.4.1", path = "utils" }
@@ -105,4 +106,4 @@ backend-registry = ["nydus-storage/backend-registry"]
 backend-s3 = ["nydus-storage/backend-s3"]
 
 [workspace]
-members = ["api", "clib", "rafs", "storage", "service", "utils"]
+members = ["api", "builder", "clib", "rafs", "storage", "service", "utils"]
