+++ builder/src/core/bootstrap.rs
@@ -7,11 +7,10 @@ use anyhow::{Context, Error, Result};
 use nydus_utils::digest::{self, RafsDigest};
 use std::ops::Deref;
 
-use crate::builder::{
-    ArtifactStorage, BlobManager, BootstrapContext, BootstrapManager, BuildContext, Tree,
-};
-use crate::metadata::layout::{RafsBlobTable, RAFS_V5_ROOT_INODE};
-use crate::metadata::{RafsSuper, RafsSuperConfig, RafsSuperFlags};
+use nydus_rafs::metadata::layout::{RafsBlobTable, RAFS_V5_ROOT_INODE};
+use nydus_rafs::metadata::{RafsSuper, RafsSuperConfig, RafsSuperFlags};
+
+use crate::{ArtifactStorage, BlobManager, BootstrapContext, BootstrapManager, BuildContext, Tree};
 
 /// RAFS bootstrap/meta builder.
 pub struct Bootstrap {
