+++ builder/src/core/v5.rs
@@ -7,18 +7,18 @@ use std::convert::TryFrom;
 use std::mem::size_of;
 
 use anyhow::{bail, Context, Result};
+use nydus_rafs::metadata::inode::InodeWrapper;
+use nydus_rafs::metadata::layout::v5::{
+    RafsV5BlobTable, RafsV5ChunkInfo, RafsV5InodeTable, RafsV5InodeWrapper, RafsV5SuperBlock,
+    RafsV5XAttrsTable,
+};
+use nydus_rafs::metadata::{RafsStore, RafsVersion};
+use nydus_rafs::RafsIoWrite;
 use nydus_utils::digest::{DigestHasher, RafsDigest};
 use nydus_utils::{div_round_up, root_tracer, timing_tracer, try_round_up_4k};
 
 use super::node::Node;
-use crate::builder::{Bootstrap, BootstrapContext, BuildContext, Tree};
-use crate::metadata::inode::InodeWrapper;
-use crate::metadata::layout::v5::{
-    RafsV5BlobTable, RafsV5ChunkInfo, RafsV5InodeTable, RafsV5InodeWrapper, RafsV5SuperBlock,
-    RafsV5XAttrsTable,
-};
-use crate::metadata::{RafsStore, RafsVersion};
-use crate::RafsIoWrite;
+use crate::{Bootstrap, BootstrapContext, BuildContext, Tree};
 
 // Filesystem may have different algorithms to calculate `i_size` for directory entries,
 // which may break "repeatable build". To support repeatable build, instead of reuse the value
