+++ service/Cargo.toml
@@ -25,7 +25,7 @@ time = { version = "0.3.14", features = ["serde-human-readable"] }
 tokio = { version = "1.24", features = ["macros"] }
 
 nydus-api = { version = "0.2.2", path = "../api"}
-nydus-rafs = { version = "0.2.2", path = "../rafs" }
+nydus-rafs = { version = "0.3.0", path = "../rafs" }
 nydus-storage = { version = "0.6.2", path = "../storage" }
 nydus-utils = { version = "0.4.1", path = "../utils" }
