+++ src/bin/nydus-image/validator.rs
@@ -9,7 +9,7 @@ use std::sync::Arc;
 
 use anyhow::{Context, Result};
 use nydus_api::ConfigV2;
-use nydus_rafs::builder::Tree;
+use nydus_builder::Tree;
 use nydus_rafs::metadata::RafsSuper;
 use nydus_storage::device::BlobInfo;
